move Loci{Set,Map} into pkgs, split out classes
each subpkg ({map,set}) has a Contig, Builder, and various Serializer
classes in addition to the Loci{Set,Map} that is the main attraction.
