+++ src/main/scala/org/hammerlab/guacamole/commands/VariantSupport.scala
@@ -25,9 +25,9 @@ import org.bdgenomics.formats.avro.Variant
 import org.hammerlab.guacamole.distributed.LociPartitionUtils
 import org.hammerlab.guacamole.distributed.LociPartitionUtils.partitionLociUniformly
 import org.hammerlab.guacamole.distributed.PileupFlatMapUtils.pileupFlatMap
-import org.hammerlab.guacamole.loci.LociSet
+import org.hammerlab.guacamole.loci.set.LociSet
 import org.hammerlab.guacamole.pileup.Pileup
-import org.hammerlab.guacamole.reads.{MappedRead, InputFilters, ReadLoadingConfigArgs}
+import org.hammerlab.guacamole.reads.{InputFilters, MappedRead, ReadLoadingConfigArgs}
 import org.hammerlab.guacamole.reference.ReferenceBroadcast
 import org.hammerlab.guacamole.{Bases, ReadSet, SparkCommand}
 import org.kohsuke.args4j.{Argument, Option => Args4jOption}
