+++ src/main/scala/org/hammerlab/guacamole/loci/SimpleRange.scala
@@ -1,0 +1,28 @@
+package org.hammerlab.guacamole.loci
+
+import scala.collection.Iterator
+
+/**
+ * A range of Longs. Inclusive on start, exclusive on end.
+ */
+case class SimpleRange(start: Long, end: Long) extends Ordered[SimpleRange] {
+  /** Iterate through elements in the range. */
+  def iterator(): Iterator[Long] = new Iterator[Long] {
+    private var i = start
+    override def hasNext: Boolean = i < end
+    override def next(): Long =
+      if (hasNext) { val result = i; i += 1; result }
+      else Iterator.empty.next()
+  }
+  /** Number of elements in the range. */
+  def length: Long = end - start
+
+  /** Comparisons between ranges. Order is DESCENDING (i.e. reversed) from by start. */
+  def compare(other: SimpleRange): Int = {
+    val diff = start - other.start
+    if (diff < 0) -1
+    else if (diff == 0) 0
+    else 1
+  }
+}
+
