+++ src/main/scala/org/hammerlab/guacamole/loci/map/Builder.scala
@@ -1,0 +1,62 @@
+package org.hammerlab.guacamole.loci.map
+
+import com.google.common.collect.{Range, TreeRangeMap}
+import org.hammerlab.guacamole.loci.set.LociSet
+
+import scala.collection.mutable
+import java.lang.{Long => JLong}
+
+import scala.collection.mutable.ArrayBuffer
+
+/** Class for building a LociMap */
+private[loci] class Builder[T] {
+  val data = new mutable.HashMap[String, mutable.ArrayBuffer[(Long, Long, T)]]()
+
+  /** Set the value at the given locus range in the LociMap under construction. */
+  def put(contig: String, start: Long, end: Long, value: T): Builder[T] = {
+    assume(end >= start)
+    if (end > start) data.get(contig) match {
+      case None              => data.put(contig, ArrayBuffer[(Long, Long, T)]((start, end, value)))
+      case Some(arrayBuffer) => arrayBuffer += ((start, end, value))
+    }
+    this
+  }
+
+  /** Set the value for all loci in the given LociSet to the specified value in the LociMap under construction. */
+  def put(loci: LociSet, value: T): Builder[T] = {
+    loci.contigs.foreach(contig => {
+      loci.onContig(contig).ranges.foreach(range => {
+        put(contig, range.start, range.end, value)
+      })
+    })
+    this
+  }
+
+  /** Build the result. */
+  def result(): LociMap[T] = {
+    LociMap[T](data.map({
+      case (contig, array) => {
+        val mutableRangeMap = TreeRangeMap.create[JLong, T]()
+        array.foreach(item => {
+          var (start, end, value) = item
+          // If there is an existing entry associated *with the same value* in the map immediately before the range
+          // we're adding, we coalesce the two ranges by setting our start to be its start.
+          val existingStart = mutableRangeMap.getEntry(start - 1)
+          if (existingStart != null && existingStart.getValue == value) {
+            assert(existingStart.getKey.lowerEndpoint < start)
+            start = existingStart.getKey.lowerEndpoint
+          }
+          // Likewise for the end of the range.
+          val existingEnd = mutableRangeMap.getEntry(end)
+          if (existingEnd != null && existingEnd.getValue == value) {
+            assert(existingEnd.getKey.upperEndpoint > end)
+            end = existingEnd.getKey.upperEndpoint
+          }
+          mutableRangeMap.put(Range.closedOpen[JLong](start, end), value)
+        })
+        contig -> Contig(contig, mutableRangeMap)
+      }
+    }).toMap)
+  }
+}
+
