+++ src/main/scala/org/hammerlab/guacamole/reads/Read.scala
@@ -29,6 +29,7 @@ import org.bdgenomics.adam.models.SequenceDictionary
 import org.bdgenomics.adam.rdd.{ADAMContext, ADAMSpecificRecordSequenceDictionaryRDDAggregator}
 import org.bdgenomics.formats.avro.AlignmentRecord
 import org.hammerlab.guacamole.Bases
+import org.hammerlab.guacamole.loci.set.{Builder => LociSetBuilder}
 import org.hammerlab.guacamole.logging.LoggingUtils.progress
 import org.seqdoop.hadoop_bam.util.SAMHeaderReader
 import org.seqdoop.hadoop_bam.{AnySAMInputFormat, SAMRecordWritable}
