+++ src/test/scala/org/hammerlab/guacamole/reads/ReadSetSuite.scala
@@ -21,7 +21,7 @@ package org.hammerlab.guacamole.reads
 import org.apache.parquet.hadoop.metadata.CompressionCodecName
 import org.bdgenomics.adam.rdd.read.AlignmentRecordRDDFunctions
 import org.bdgenomics.adam.rdd.{ADAMContext, ADAMSaveAnyArgs}
-import org.hammerlab.guacamole.loci.LociSet
+import org.hammerlab.guacamole.loci.set.LociSet
 import org.hammerlab.guacamole.util.{GuacFunSuite, TestUtil}
 
 class ReadSetSuite extends GuacFunSuite {
