+++ src/test/scala/org/hammerlab/guacamole/util/TestUtil.scala
@@ -19,8 +19,7 @@
 package org.hammerlab.guacamole.util
 
 import java.io.{File, FileNotFoundException}
-import java.nio.file.{Files, Path}
-import java.util.UUID
+import java.nio.file.Files
 
 import com.esotericsoftware.kryo.Kryo
 import com.twitter.chill.{IKryoRegistrar, KryoInstantiator, KryoPool}
@@ -28,10 +27,9 @@ import htsjdk.samtools.TextCigarCodec
 import org.apache.commons.io.FileUtils
 import org.apache.spark.SparkContext
 import org.hammerlab.guacamole.kryo.GuacamoleKryoRegistrator
-import org.hammerlab.guacamole.loci.LociSet
+import org.hammerlab.guacamole.loci.set.LociSet
 import org.hammerlab.guacamole.pileup.Pileup
-import org.hammerlab.guacamole.reads.InputFilters
-import org.hammerlab.guacamole.reads._
+import org.hammerlab.guacamole.reads.{InputFilters, MappedRead, MateAlignmentProperties, PairedRead, Read, ReadLoadingConfig}
 import org.hammerlab.guacamole.reference.ReferenceBroadcast.MapBackedReferenceSequence
 import org.hammerlab.guacamole.reference.{ContigSequence, ReferenceBroadcast}
 import org.hammerlab.guacamole.{Bases, ReadSet}
